lifted albums state to dashboard
