+++ client/src/Components/Featuring/TrackCard/TrackCard.jsx
@@ -17,6 +17,11 @@ const TrackCard = ({ track }) => {
     console.log("clicked");
   };
 
+  const handlePlay = (e) => {
+    e.stopPropagation();
+    console.log("Play Clicked");
+  };
+
   return (
     <>
       <button
@@ -33,7 +38,11 @@ const TrackCard = ({ track }) => {
             alt={track.album_name}
           ></img>
           {isHovering ? (
-            <img className={styles["play-button"]} src={playButton}></img>
+            <img
+              className={styles["play-button"]}
+              onClick={handlePlay}
+              src={playButton}
+            ></img>
           ) : (
             <img
               className={styles["play-button-hidden"]}
